lib: matcher: widen matcher type bitmask to 64 bits
With 30 matcher types defined, the uint32_t bitmask guard only has 2
slots remaining. Widen the static_assert to uint64_t to allow the enum
to grow beyond 32 values, and widen the mask variable in bf_set_new()
to uint64_t to prevent silent truncation of BF_FLAG() results.
BF_FLAG() already uses 1ULL, so no other changes are needed.

Closes: https://github.com/facebook/bpfilter/issues/443

+++ src/libbpfilter/set.c
@@ -26,7 +26,7 @@ int bf_set_new(struct bf_set **set, const char *name, enum bf_matcher_type *key,
                size_t n_comps)
 {
     _free_bf_set_ struct bf_set *_set = NULL;
-    uint32_t mask = 0;
+    uint64_t mask = 0;
 
     assert(set);
     assert(key);
